+++ src/main/java/kr/modusplant/modules/auth/normal/app/service/NormalSignUpApplicationService.java
@@ -1,17 +1,15 @@
 package kr.modusplant.modules.auth.normal.app.service;
 
-import kr.modusplant.domains.member.app.http.request.SiteMemberAuthInsertRequest;
 import kr.modusplant.domains.member.app.http.response.SiteMemberResponse;
 import kr.modusplant.domains.member.app.service.SiteMemberApplicationService;
 import kr.modusplant.domains.member.app.service.SiteMemberAuthApplicationService;
 import kr.modusplant.domains.member.app.service.SiteMemberTermApplicationService;
-import kr.modusplant.domains.member.domain.model.SiteMember;
 import kr.modusplant.domains.term.app.http.response.TermResponse;
 import kr.modusplant.domains.term.app.service.TermApplicationService;
 import kr.modusplant.modules.auth.normal.app.http.request.NormalSignUpRequest;
-import kr.modusplant.modules.auth.normal.mapper.domain.SiteMemberAuthDomainMapper;
-import kr.modusplant.modules.auth.normal.mapper.domain.SiteMemberDomainMapper;
-import kr.modusplant.modules.auth.normal.mapper.domain.SiteMemberTermDomainMapper;
+import kr.modusplant.modules.auth.normal.mapper.NormalSignUpMemberAppDomainMapper;
+import kr.modusplant.modules.auth.normal.mapper.NormalSignupAuthAppDomainMapper;
+import kr.modusplant.modules.auth.normal.mapper.NormalSignupTermAppDomainMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,9 +25,9 @@ public class NormalSignUpApplicationService {
     private final SiteMemberApplicationService siteMemberApplicationService;
     private final SiteMemberAuthApplicationService siteMemberAuthApplicationService;
     private final SiteMemberTermApplicationService siteMemberTermApplicationService;
-    private final SiteMemberDomainMapper siteMemberDomainMapper;
-    private final SiteMemberAuthDomainMapper siteMemberAuthDomainMapper;
-    private final SiteMemberTermDomainMapper siteMemberTermDomainMapper;
+    private final NormalSignUpMemberAppDomainMapper normalSignUpMemberAppDomainMapper;
+    private final NormalSignupAuthAppDomainMapper normalSignupAuthAppDomainMapper;
+    private final NormalSignupTermAppDomainMapper normalSignupTermAppDomainMapper;
 
     public List<TermResponse> getAllTerms() {
         return termApplicationService.getAll();
@@ -51,9 +49,9 @@ public class NormalSignUpApplicationService {
 
     @Transactional
     public void insertMember(NormalSignUpRequest request) {
-        SiteMemberResponse savedMember = siteMemberApplicationService.insert(siteMemberDomainMapper.toSiteMemberInsertRequest(request));
-        siteMemberAuthApplicationService.insert(siteMemberAuthDomainMapper.toSiteMemberAuthInsertRequest(request, savedMember));
-        siteMemberTermApplicationService.insert(siteMemberTermDomainMapper.toSiteMemberTermInsertRequest(request, savedMember));
+        SiteMemberResponse savedMember = siteMemberApplicationService.insert(normalSignUpMemberAppDomainMapper.toSiteMemberInsertRequest(request));
+        siteMemberAuthApplicationService.insert(normalSignupAuthAppDomainMapper.toSiteMemberAuthInsertRequest(request, savedMember));
+        siteMemberTermApplicationService.insert(normalSignupTermAppDomainMapper.toSiteMemberTermInsertRequest(request, savedMember));
     }
 
     private Map<String, Object> createTermMap(TermResponse term) {
