+++ src/main/java/kr/modusplant/modules/auth/normal/mapper/NormalSignupTermAppDomainMapper.java
@@ -1,18 +1,13 @@
-package kr.modusplant.modules.auth.normal.mapper.domain;
+package kr.modusplant.modules.auth.normal.mapper;
 
 import kr.modusplant.domains.member.app.http.request.SiteMemberTermInsertRequest;
 import kr.modusplant.domains.member.app.http.response.SiteMemberResponse;
-import kr.modusplant.domains.member.domain.model.SiteMember;
-import kr.modusplant.domains.member.domain.model.SiteMemberTerm;
 import kr.modusplant.modules.auth.normal.app.http.request.NormalSignUpRequest;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import org.mapstruct.NullValuePropertyMappingStrategy;
-import org.mapstruct.ReportingPolicy;
 
-@Mapper(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.SET_TO_NULL,
-        unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface SiteMemberTermDomainMapper {
+@Mapper
+public interface NormalSignupTermAppDomainMapper {
 
     @Mapping(target = "uuid", source = "memberResponse.uuid")
     SiteMemberTermInsertRequest toSiteMemberTermInsertRequest(NormalSignUpRequest request, SiteMemberResponse memberResponse);
